PR-14: extract sni/ (8-file directory module)
Largest single-PR extraction so far (~1015 LOC). Behaviour-preserving.

- sni/mod.rs: SniControl variant enum, SniControlMode,
  SniRuntimeTransitionPlan, SniRuntimeWakeReason, mode-resolution
  and transition-planning helpers. pub(crate) use re-exports all
  submodules per D24.
- sni/settings.rs: DconfBackend trait + ShellDconfBackend,
  SniSettingsStore, dconf helpers.
- sni/state.rs: MenuRefresh + SniIndicatorState.
- sni/indicator.rs: SniIndicator + impl Tray, start_sni_indicator,
  SniIndicatorRuntimeHandle + Drop. Test-only ACTIVE_SNI_WATCHER_TASKS
  + SniWatcherTaskGuard + sni_watcher_task_count under cfg(test).
- sni/control_{local,dbus}.rs: SniLocalControl and SniDbusControl
  structs (fields pub(crate) so control_ops can pattern-match).
- sni/control_ops.rs: SniControlOps trait + impl for SniControl.
- sni/guard.rs: SniGuard + build_sni_control_for_mode +
  SNI_RUNTIME_RETRY_INTERVAL.

main.rs: 1979 -> 958 LOC. (Broke below 1000 LOC.) Removed ksni and
noto_sans_mono_bitmap top-level imports.

Verification: cargo build clean; cargo test --bin kanata-switcher
261 passed / 0 failed.
