+++ src/daemon/main.rs
@@ -1,10 +1,5 @@
 use clap::{ArgMatches, CommandFactory, FromArgMatches, Parser, ValueEnum};
 use futures_util::StreamExt;
-use ksni::menu::{CheckmarkItem, StandardItem};
-use ksni::{Icon as SniIcon, MenuItem, Status as SniStatus, ToolTip, Tray, TrayService};
-use noto_sans_mono_bitmap::{
-    FontWeight, RasterHeight, RasterizedChar, get_raster, get_raster_width,
-};
 use std::collections::HashMap;
 use std::env;
 use std::fs;
@@ -12,11 +7,8 @@ use std::os::fd::AsFd;
 use std::os::unix::net::UnixStream;
 use std::path::{Path, PathBuf};
 use std::process::Command;
-#[cfg(test)]
-use std::sync::atomic::AtomicUsize;
 use std::sync::atomic::{AtomicBool, AtomicU64, Ordering};
 use std::sync::{Arc, Mutex};
-use std::thread;
 use std::time::Duration;
 use tokio::sync::{Mutex as TokioMutex, mpsc, oneshot, watch};
 use uuid::Uuid;
@@ -40,6 +32,7 @@ mod lifecycle;
 mod display_override;
 mod supervisor;
 mod backends;
+mod sni;
 
 use constants::*;
 use errors::DynError;
@@ -62,1027 +55,13 @@ use display_override::*;
 use supervisor::*;
 use backends::*;
 use backends::wayland::*;
+use sni::*;
 
 #[cfg(test)]
 #[allow(unused_imports)]
-pub(crate) use crate::{constants::*, errors::*, environ::*, dbus_naming::*, config::*, focus::*, args::*, autostart::*, broadcasters::*, kanata::*, control::*, control::client::*, control::server::*, control::persistent::*, pause::*, focus_pipeline::*, lifecycle::*, lifecycle::logind::*, lifecycle::startup::*, display_override::*, supervisor::*, supervisor::capabilities::*, backends::*, backends::gnome::*, backends::kde::*, backends::kde::script::*, backends::kde::probe::*, backends::wayland::*};
+pub(crate) use crate::{constants::*, errors::*, environ::*, dbus_naming::*, config::*, focus::*, args::*, autostart::*, broadcasters::*, kanata::*, control::*, control::client::*, control::server::*, control::persistent::*, pause::*, focus_pipeline::*, lifecycle::*, lifecycle::logind::*, lifecycle::startup::*, display_override::*, supervisor::*, supervisor::capabilities::*, backends::*, backends::gnome::*, backends::kde::*, backends::kde::script::*, backends::kde::probe::*, backends::wayland::*, sni::*, sni::settings::*, sni::state::*, sni::indicator::*, sni::control_local::*, sni::control_dbus::*, sni::control_ops::*, sni::guard::*};
 
 
-// === SNI Indicator ===
-
-pub(crate) const SNI_DEFAULT_SHOW_FOCUS_ONLY: bool = true;
-const SNI_FONT_WEIGHT: FontWeight = FontWeight::Regular;
-const SNI_RASTER_HEIGHT: RasterHeight = RasterHeight::Size32;
-const SNI_GLYPH_WIDTH: usize = get_raster_width(SNI_FONT_WEIGHT, SNI_RASTER_HEIGHT);
-const SNI_GLYPH_HEIGHT: usize = SNI_RASTER_HEIGHT.val();
-const SNI_GLYPH_GAP: usize = 4;
-const SNI_ICON_HEIGHT: usize = SNI_GLYPH_HEIGHT;
-const SNI_COLOR_LAYER: [u8; 4] = [255, 255, 255, 255];
-const SNI_COLOR_VK: [u8; 4] = [255, 0, 255, 255];
-const SNI_MAX_VK_COUNT_DIGIT: usize = 9;
-const SNI_MIN_MULTI_VK_COUNT: usize = 2;
-const SNI_INDICATOR_ID: &str = "kanata-switcher";
-
-trait DconfBackend: Send + Sync {
-    fn get_bool(&self, key: &str) -> Result<bool, String>;
-    fn set_bool(&self, key: &str, value: bool) -> Result<(), String>;
-}
-
-struct ShellDconfBackend;
-
-impl DconfBackend for ShellDconfBackend {
-    fn get_bool(&self, key: &str) -> Result<bool, String> {
-        dconf_get_bool(key)
-    }
-
-    fn set_bool(&self, key: &str, value: bool) -> Result<(), String> {
-        dconf_set_bool(key, value)
-    }
-}
-
-pub(crate) struct SniSettingsStore {
-    available: bool,
-    backend: Box<dyn DconfBackend>,
-}
-
-impl SniSettingsStore {
-    fn new() -> Self {
-        Self {
-            available: true,
-            backend: Box::new(ShellDconfBackend),
-        }
-    }
-
-    #[cfg(test)]
-    fn with_backend(backend: Box<dyn DconfBackend>) -> Self {
-        Self {
-            available: true,
-            backend,
-        }
-    }
-
-    #[cfg(test)]
-    fn disabled() -> Self {
-        Self {
-            available: false,
-            backend: Box::new(ShellDconfBackend),
-        }
-    }
-
-    fn read_focus_only(&mut self) -> Option<bool> {
-        if !self.available {
-            return None;
-        }
-        match self.backend.get_bool(DCONF_FOCUS_ONLY_KEY) {
-            Ok(value) => Some(value),
-            Err(error) => {
-                if is_dconf_unavailable(&error) {
-                    self.available = false;
-                }
-                eprintln!("[SNI] dconf read failed: {}", error);
-                None
-            }
-        }
-    }
-
-    fn write_focus_only(&mut self, value: bool) {
-        if !self.available {
-            return;
-        }
-        if let Err(error) = self.backend.set_bool(DCONF_FOCUS_ONLY_KEY, value) {
-            if is_dconf_unavailable(&error) {
-                self.available = false;
-            }
-            eprintln!("[SNI] dconf write failed: {}", error);
-        }
-    }
-}
-
-struct MenuRefresh {
-    sender: watch::Sender<u64>,
-    version: u64,
-}
-
-impl MenuRefresh {
-    fn new() -> (Self, watch::Receiver<u64>) {
-        let (sender, receiver) = watch::channel(0u64);
-        (Self { sender, version: 0 }, receiver)
-    }
-
-    fn notify(&mut self) {
-        self.version += 1;
-        self.sender.send_replace(self.version);
-    }
-}
-
-#[derive(Clone, Debug)]
-struct SniIndicatorState {
-    last_status: StatusSnapshot,
-    focus_status: StatusSnapshot,
-    paused: bool,
-    show_focus_only: bool,
-    menu_revision: u64,
-}
-
-impl SniIndicatorState {
-    fn new(initial: StatusSnapshot, show_focus_only: bool) -> Self {
-        Self {
-            last_status: initial.clone(),
-            focus_status: initial,
-            paused: false,
-            show_focus_only,
-            menu_revision: 0,
-        }
-    }
-
-    fn update_status(&mut self, snapshot: StatusSnapshot) {
-        if snapshot.layer_source == LayerSource::Focus {
-            self.focus_status = snapshot.clone();
-        }
-        self.last_status = snapshot;
-    }
-
-    fn set_paused(&mut self, paused: bool) {
-        self.paused = paused;
-    }
-
-    fn toggle_focus_only(&mut self) {
-        self.show_focus_only = !self.show_focus_only;
-    }
-
-    fn focus_only_enabled(&self) -> bool {
-        self.show_focus_only
-    }
-
-    fn bump_menu_revision(&mut self) {
-        self.menu_revision = self.menu_revision.wrapping_add(1);
-    }
-
-    fn display_status(&self) -> StatusSnapshot {
-        if self.paused {
-            return self.last_status.clone();
-        }
-        if self.show_focus_only {
-            return self.focus_status.clone();
-        }
-        self.last_status.clone()
-    }
-}
-
-#[derive(Clone)]
-struct SniLocalControl {
-    runtime_handle: tokio::runtime::Handle,
-    kanata: KanataClient,
-    handler: Arc<Mutex<FocusHandler>>,
-    status_broadcaster: StatusBroadcaster,
-    pause_broadcaster: PauseBroadcaster,
-    restart_handle: RestartHandle,
-    shutdown_handle: ShutdownHandle,
-    unpause_context: UnpauseContext,
-}
-
-#[derive(Clone)]
-struct SniDbusControl {
-    runtime_handle: tokio::runtime::Handle,
-    connection: Connection,
-    restart_handle: RestartHandle,
-    shutdown_handle: ShutdownHandle,
-    /// Per-instance daemon bus name to target with DBus control calls.
-    daemon_bus_name: String,
-}
-
-#[derive(Clone)]
-enum SniControl {
-    Local(SniLocalControl),
-    Dbus(SniDbusControl),
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum SniControlMode {
-    Local,
-    Dbus,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum SniRuntimeTransitionPlan {
-    Keep,
-    Stop,
-    Start(SniControlMode),
-    Restart(SniControlMode),
-}
-
-fn sni_control_mode_for_environment(env: Environment) -> Option<SniControlMode> {
-    match env {
-        Environment::Wayland | Environment::X11 => Some(SniControlMode::Local),
-        Environment::Kde => Some(SniControlMode::Dbus),
-        Environment::Gnome | Environment::LinuxConsoleWithLogind | Environment::Unknown => None,
-    }
-}
-
-fn plan_sni_runtime_transition(
-    active_mode: Option<SniControlMode>,
-    active_env: Option<Environment>,
-    desired_env: Environment,
-) -> SniRuntimeTransitionPlan {
-    let desired_mode = sni_control_mode_for_environment(desired_env);
-    match (active_mode, desired_mode) {
-        (None, None) => SniRuntimeTransitionPlan::Keep,
-        (Some(_), None) => SniRuntimeTransitionPlan::Stop,
-        (None, Some(mode)) => SniRuntimeTransitionPlan::Start(mode),
-        (Some(current_mode), Some(next_mode)) => {
-            if current_mode != next_mode || active_env != Some(desired_env) {
-                SniRuntimeTransitionPlan::Restart(next_mode)
-            } else {
-                SniRuntimeTransitionPlan::Keep
-            }
-        }
-    }
-}
-
-const SNI_RUNTIME_RETRY_INTERVAL: Duration = Duration::from_secs(1);
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum SniRuntimeWakeReason {
-    EnvironmentChanged,
-    Retry,
-    ChannelClosed,
-}
-
-async fn wait_for_sni_runtime_wake_with_delay(
-    env_receiver: &mut watch::Receiver<Environment>,
-    should_retry_start: bool,
-    retry_delay: Duration,
-) -> SniRuntimeWakeReason {
-    if !should_retry_start {
-        return match env_receiver.changed().await {
-            Ok(_) => SniRuntimeWakeReason::EnvironmentChanged,
-            Err(_) => SniRuntimeWakeReason::ChannelClosed,
-        };
-    }
-
-    tokio::select! {
-        changed = env_receiver.changed() => {
-            match changed {
-                Ok(_) => SniRuntimeWakeReason::EnvironmentChanged,
-                Err(_) => SniRuntimeWakeReason::ChannelClosed,
-            }
-        }
-        _ = tokio::time::sleep(retry_delay) => SniRuntimeWakeReason::Retry,
-    }
-}
-
-trait SniControlOps: Send + Sync {
-    fn restart(&self);
-    fn pause(&self);
-    fn unpause(&self);
-    fn quit(&self);
-}
-
-impl SniControlOps for SniControl {
-    fn restart(&self) {
-        println!("[SNI] Restart requested");
-        match self {
-            SniControl::Local(control) => {
-                control.restart_handle.request();
-            }
-            SniControl::Dbus(control) => {
-                control.runtime_handle.block_on(async {
-                    if let Err(error) = send_control_command_with_connection(
-                        &control.connection,
-                        &control.daemon_bus_name,
-                        ControlCommand::Restart,
-                    )
-                    .await
-                    {
-                        eprintln!("[SNI] Failed to send restart: {}", error);
-                    }
-                });
-                control.restart_handle.request();
-            }
-        }
-    }
-
-    fn pause(&self) {
-        println!("[SNI] Pause requested");
-        match self {
-            SniControl::Local(control) => {
-                pause_daemon(
-                    &control.pause_broadcaster,
-                    &control.handler,
-                    &control.status_broadcaster,
-                    &control.kanata,
-                    &control.runtime_handle,
-                    "via SNI",
-                );
-            }
-            SniControl::Dbus(control) => {
-                control.runtime_handle.block_on(async {
-                    if let Err(error) = send_control_command_with_connection(
-                        &control.connection,
-                        &control.daemon_bus_name,
-                        ControlCommand::Pause,
-                    )
-                    .await
-                    {
-                        eprintln!("[SNI] Failed to send pause: {}", error);
-                    }
-                });
-            }
-        }
-    }
-
-    fn unpause(&self) {
-        println!("[SNI] Unpause requested");
-        match self {
-            SniControl::Local(control) => {
-                let context = control.unpause_context.clone();
-                unpause_daemon(
-                    context.env,
-                    context.connection,
-                    context.is_kde6,
-                    &control.pause_broadcaster,
-                    &control.handler,
-                    &control.status_broadcaster,
-                    &control.kanata,
-                    &control.runtime_handle,
-                    "via SNI",
-                );
-            }
-            SniControl::Dbus(control) => {
-                control.runtime_handle.block_on(async {
-                    if let Err(error) = send_control_command_with_connection(
-                        &control.connection,
-                        &control.daemon_bus_name,
-                        ControlCommand::Unpause,
-                    )
-                    .await
-                    {
-                        eprintln!("[SNI] Failed to send unpause: {}", error);
-                    }
-                });
-            }
-        }
-    }
-
-    fn quit(&self) {
-        println!("[SNI] Quit requested");
-        match self {
-            SniControl::Local(control) => {
-                control.shutdown_handle.request();
-            }
-            SniControl::Dbus(control) => {
-                control.shutdown_handle.request();
-            }
-        }
-    }
-}
-
-struct SniIndicator {
-    state: SniIndicatorState,
-    control: Arc<dyn SniControlOps>,
-    settings: SniSettingsStore,
-    menu_refresh: MenuRefresh,
-}
-
-impl SniIndicator {
-    fn update_status(&mut self, snapshot: StatusSnapshot) {
-        self.state.update_status(snapshot);
-    }
-
-    fn set_paused(&mut self, paused: bool) {
-        self.state.set_paused(paused);
-    }
-
-    fn toggle_focus_only(&mut self) {
-        self.state.toggle_focus_only();
-        let show_focus_only = self.state.focus_only_enabled();
-        self.settings.write_focus_only(show_focus_only);
-        self.menu_refresh.notify();
-    }
-
-    fn request_pause(&self) {
-        if self.state.paused {
-            self.control.unpause();
-        } else {
-            self.control.pause();
-        }
-    }
-
-    fn request_restart(&self) {
-        self.control.restart();
-    }
-
-    fn request_quit(&self) {
-        self.control.quit();
-    }
-
-    fn format_layer_letter(layer_name: &str) -> String {
-        let trimmed = layer_name.trim();
-        if trimmed.is_empty() {
-            return "?".to_string();
-        }
-        trimmed
-            .chars()
-            .next()
-            .map(|c| c.to_uppercase().to_string())
-            .unwrap_or_else(|| "?".to_string())
-    }
-
-    fn format_virtual_keys(virtual_keys: &[String]) -> String {
-        let count = virtual_keys.len();
-        if count == 0 {
-            return String::new();
-        }
-        if count == 1 {
-            let name = virtual_keys[0].trim();
-            if name.is_empty() {
-                return String::new();
-            }
-            return name
-                .chars()
-                .next()
-                .map(|c| c.to_uppercase().to_string())
-                .unwrap_or_default();
-        }
-        if count < SNI_MIN_MULTI_VK_COUNT {
-            return String::new();
-        }
-        if count > SNI_MAX_VK_COUNT_DIGIT {
-            return format!("{}+", SNI_MAX_VK_COUNT_DIGIT);
-        }
-        count.to_string()
-    }
-
-    fn glyph_for_char(ch: char) -> RasterizedChar {
-        get_raster(ch, SNI_FONT_WEIGHT, SNI_RASTER_HEIGHT)
-            .or_else(|| get_raster('?', SNI_FONT_WEIGHT, SNI_RASTER_HEIGHT))
-            .expect("SNI glyph lookup failed")
-    }
-
-    fn draw_glyph(
-        buffer: &mut [u8],
-        width: usize,
-        height: usize,
-        x: usize,
-        y: usize,
-        glyph: &RasterizedChar,
-        color: [u8; 4],
-    ) {
-        for (row_index, row) in glyph.raster().iter().enumerate() {
-            let dest_y = y + row_index;
-            if dest_y >= height {
-                continue;
-            }
-            for (col_index, intensity) in row.iter().enumerate() {
-                let dest_x = x + col_index;
-                if dest_x >= width {
-                    continue;
-                }
-                if *intensity == 0 {
-                    continue;
-                }
-                let offset = (dest_y * width + dest_x) * 4;
-                let alpha = (u16::from(color[0]) * u16::from(*intensity) / 255) as u8;
-                let red = (u16::from(color[1]) * u16::from(*intensity) / 255) as u8;
-                let green = (u16::from(color[2]) * u16::from(*intensity) / 255) as u8;
-                let blue = (u16::from(color[3]) * u16::from(*intensity) / 255) as u8;
-                buffer[offset] = alpha;
-                buffer[offset + 1] = red;
-                buffer[offset + 2] = green;
-                buffer[offset + 3] = blue;
-            }
-        }
-    }
-
-    fn text_width(text: &str) -> usize {
-        SNI_GLYPH_WIDTH * text.chars().count()
-    }
-
-    fn draw_text(
-        buffer: &mut [u8],
-        width: usize,
-        height: usize,
-        x: usize,
-        y: usize,
-        text: &str,
-        color: [u8; 4],
-    ) -> usize {
-        let mut cursor_x = x;
-        for ch in text.chars() {
-            let glyph = Self::glyph_for_char(ch);
-            Self::draw_glyph(buffer, width, height, cursor_x, y, &glyph, color);
-            cursor_x += glyph.width();
-        }
-        cursor_x - x
-    }
-
-    fn render_icon(layer_text: &str, vk_text: &str) -> SniIcon {
-        let layer_width = Self::text_width(layer_text);
-        let vk_width = Self::text_width(vk_text);
-        let gap = if vk_text.is_empty() { 0 } else { SNI_GLYPH_GAP };
-        let icon_width = layer_width + gap + vk_width;
-        let mut buffer = vec![0u8; icon_width * SNI_ICON_HEIGHT * 4];
-        let glyph_y = (SNI_ICON_HEIGHT - SNI_GLYPH_HEIGHT) / 2;
-        let layer_x = 0;
-        let vk_x = layer_x + layer_width + gap;
-
-        Self::draw_text(
-            &mut buffer,
-            icon_width,
-            SNI_ICON_HEIGHT,
-            layer_x,
-            glyph_y,
-            layer_text,
-            SNI_COLOR_LAYER,
-        );
-        if !vk_text.is_empty() {
-            Self::draw_text(
-                &mut buffer,
-                icon_width,
-                SNI_ICON_HEIGHT,
-                vk_x,
-                glyph_y,
-                vk_text,
-                SNI_COLOR_VK,
-            );
-        }
-
-        SniIcon {
-            width: icon_width as i32,
-            height: SNI_ICON_HEIGHT as i32,
-            data: buffer,
-        }
-    }
-
-    fn display_strings(&self) -> (String, String) {
-        let status = self.state.display_status();
-        let layer_text = Self::format_layer_letter(&status.layer);
-        let vk_text = Self::format_virtual_keys(&status.virtual_keys);
-        (layer_text, vk_text)
-    }
-
-    fn tooltip_text(&self) -> String {
-        let status = self.state.display_status();
-        if status.virtual_keys.is_empty() {
-            return format!("Layer: {}", status.layer);
-        }
-        format!(
-            "Layer: {}\nVirtual keys: {}",
-            status.layer,
-            status.virtual_keys.join(", ")
-        )
-    }
-
-    fn title_text(&self) -> String {
-        "Kanata Switcher".to_string()
-    }
-}
-
-impl Tray for SniIndicator {
-    fn id(&self) -> String {
-        SNI_INDICATOR_ID.to_string()
-    }
-
-    fn title(&self) -> String {
-        self.title_text()
-    }
-
-    fn status(&self) -> SniStatus {
-        SniStatus::Active
-    }
-
-    fn icon_pixmap(&self) -> Vec<SniIcon> {
-        let (layer_text, vk_text) = self.display_strings();
-        vec![Self::render_icon(&layer_text, &vk_text)]
-    }
-
-    fn tool_tip(&self) -> ToolTip {
-        ToolTip {
-            title: self.title_text(),
-            description: self.tooltip_text(),
-            ..ToolTip::default()
-        }
-    }
-
-    fn menu(&self) -> Vec<MenuItem<Self>> {
-        vec![
-            MenuItem::Checkmark(CheckmarkItem {
-                label: "Pause".to_string(),
-                checked: self.state.paused,
-                activate: Box::new(|this| {
-                    this.request_pause();
-                }),
-                ..CheckmarkItem::default()
-            }),
-            MenuItem::Checkmark(CheckmarkItem {
-                label: "Show app layer only".to_string(),
-                checked: self.state.show_focus_only,
-                activate: Box::new(|this| {
-                    this.toggle_focus_only();
-                }),
-                ..CheckmarkItem::default()
-            }),
-            MenuItem::Separator,
-            MenuItem::Standard(StandardItem {
-                label: "Restart".to_string(),
-                activate: Box::new(|this| {
-                    this.request_restart();
-                }),
-                ..StandardItem::default()
-            }),
-            MenuItem::Standard(StandardItem {
-                label: "Quit".to_string(),
-                activate: Box::new(|this| {
-                    this.request_quit();
-                }),
-                ..StandardItem::default()
-            }),
-        ]
-    }
-
-    fn watcher_online(&self) {
-        println!("[SNI] StatusNotifierWatcher online");
-    }
-
-    fn watcher_offine(&self) -> bool {
-        eprintln!("[SNI] StatusNotifierWatcher offline");
-        true
-    }
-}
-
-
-fn start_sni_indicator(
-    control: SniControl,
-    status_broadcaster: StatusBroadcaster,
-    pause_broadcaster: PauseBroadcaster,
-    indicator_focus_only: Option<TrayFocusOnly>,
-) -> Option<SniIndicatorRuntimeHandle> {
-    println!("[SNI] Starting StatusNotifier indicator");
-    let initial_status = status_broadcaster.snapshot();
-    let mut settings = SniSettingsStore::new();
-    let show_focus_only = resolve_sni_focus_only(indicator_focus_only, &mut settings);
-    let (menu_refresh, mut menu_refresh_receiver) = MenuRefresh::new();
-    let control_handle: Arc<dyn SniControlOps> = Arc::new(control);
-    let indicator = SniIndicator {
-        state: SniIndicatorState::new(initial_status, show_focus_only),
-        control: control_handle,
-        settings,
-        menu_refresh,
-    };
-    let service = TrayService::new(indicator);
-    let handle = service.handle();
-
-    let pause_initial = pause_broadcaster.is_paused();
-    handle.update(|state| state.set_paused(pause_initial));
-
-    let status_handle = handle.clone();
-    let mut status_receiver = status_broadcaster.subscribe();
-    let status_watch_task = tokio::spawn(async move {
-        #[cfg(test)]
-        let _watcher_guard = SniWatcherTaskGuard::new();
-        loop {
-            if status_receiver.changed().await.is_err() {
-                break;
-            }
-            let snapshot = status_receiver.borrow().clone();
-            status_handle.update(|state| state.update_status(snapshot));
-        }
-    });
-
-    let pause_handle = handle.clone();
-    let mut pause_receiver = pause_broadcaster.subscribe();
-    let pause_watch_task = tokio::spawn(async move {
-        #[cfg(test)]
-        let _watcher_guard = SniWatcherTaskGuard::new();
-        loop {
-            if pause_receiver.changed().await.is_err() {
-                break;
-            }
-            let paused = *pause_receiver.borrow();
-            pause_handle.update(|state| state.set_paused(paused));
-        }
-    });
-
-    let menu_handle = handle.clone();
-    let menu_watch_task = tokio::spawn(async move {
-        #[cfg(test)]
-        let _watcher_guard = SniWatcherTaskGuard::new();
-        loop {
-            if menu_refresh_receiver.changed().await.is_err() {
-                break;
-            }
-            menu_handle.update(|state| state.state.bump_menu_revision());
-        }
-    });
-
-    thread::spawn(move || match service.run() {
-        Ok(()) => println!("[SNI] Indicator stopped"),
-        Err(error) => eprintln!("[SNI] Failed to run indicator: {}", error),
-    });
-
-    Some(SniIndicatorRuntimeHandle {
-        handle,
-        status_watch_task,
-        pause_watch_task,
-        menu_watch_task,
-    })
-}
-
-async fn build_sni_control_for_mode(
-    mode: SniControlMode,
-    runtime_handle: tokio::runtime::Handle,
-    kanata: KanataClient,
-    handler: Arc<Mutex<FocusHandler>>,
-    status_broadcaster: StatusBroadcaster,
-    pause_broadcaster: PauseBroadcaster,
-    restart_handle: RestartHandle,
-    shutdown_handle: ShutdownHandle,
-    control_environment: Environment,
-    daemon_bus_name: String,
-) -> Option<SniControl> {
-    match mode {
-        SniControlMode::Local => Some(SniControl::Local(SniLocalControl {
-            runtime_handle,
-            kanata,
-            handler,
-            status_broadcaster,
-            pause_broadcaster,
-            restart_handle,
-            shutdown_handle,
-            unpause_context: local_sni_unpause_context(control_environment),
-        })),
-        SniControlMode::Dbus => match Connection::session().await {
-            Ok(connection) => Some(SniControl::Dbus(SniDbusControl {
-                runtime_handle,
-                connection,
-                restart_handle,
-                shutdown_handle,
-                daemon_bus_name,
-            })),
-            Err(error) => {
-                eprintln!("[SNI] Failed to connect to session bus: {}", error);
-                None
-            }
-        },
-    }
-}
-
-struct SniIndicatorRuntimeHandle {
-    handle: ksni::Handle<SniIndicator>,
-    status_watch_task: tokio::task::JoinHandle<()>,
-    pause_watch_task: tokio::task::JoinHandle<()>,
-    menu_watch_task: tokio::task::JoinHandle<()>,
-}
-
-impl SniIndicatorRuntimeHandle {
-    fn shutdown(&self) {
-        self.status_watch_task.abort();
-        self.pause_watch_task.abort();
-        self.menu_watch_task.abort();
-        self.handle.shutdown();
-    }
-}
-
-impl Drop for SniIndicatorRuntimeHandle {
-    fn drop(&mut self) {
-        self.shutdown();
-    }
-}
-
-#[cfg(test)]
-static ACTIVE_SNI_WATCHER_TASKS: AtomicUsize = AtomicUsize::new(0);
-
-#[cfg(test)]
-struct SniWatcherTaskGuard;
-
-#[cfg(test)]
-impl SniWatcherTaskGuard {
-    fn new() -> Self {
-        ACTIVE_SNI_WATCHER_TASKS.fetch_add(1, Ordering::SeqCst);
-        Self
-    }
-}
-
-#[cfg(test)]
-impl Drop for SniWatcherTaskGuard {
-    fn drop(&mut self) {
-        ACTIVE_SNI_WATCHER_TASKS.fetch_sub(1, Ordering::SeqCst);
-    }
-}
-
-#[cfg(test)]
-fn sni_watcher_task_count() -> usize {
-    ACTIVE_SNI_WATCHER_TASKS.load(Ordering::SeqCst)
-}
-
-struct SniGuard {
-    handle: Arc<Mutex<Option<SniIndicatorRuntimeHandle>>>,
-    task: Option<tokio::task::JoinHandle<()>>,
-}
-
-impl SniGuard {
-    fn disabled() -> Self {
-        Self {
-            handle: Arc::new(Mutex::new(None)),
-            task: None,
-        }
-    }
-
-    fn runtime_managed(
-        runtime_environment: RuntimeEnvironmentBroadcaster,
-        runtime_handle: tokio::runtime::Handle,
-        kanata: KanataClient,
-        handler: Arc<Mutex<FocusHandler>>,
-        status_broadcaster: StatusBroadcaster,
-        pause_broadcaster: PauseBroadcaster,
-        restart_handle: RestartHandle,
-        shutdown_handle: ShutdownHandle,
-        indicator_focus_only: Option<TrayFocusOnly>,
-        daemon_bus_name: String,
-    ) -> Self {
-        Self::runtime_managed_with_builder(
-            runtime_environment,
-            runtime_handle,
-            kanata,
-            handler,
-            status_broadcaster,
-            pause_broadcaster,
-            restart_handle,
-            shutdown_handle,
-            indicator_focus_only,
-            SNI_RUNTIME_RETRY_INTERVAL,
-            build_sni_control_for_mode,
-            daemon_bus_name,
-        )
-    }
-
-    fn runtime_managed_with_builder<B, BFut>(
-        runtime_environment: RuntimeEnvironmentBroadcaster,
-        runtime_handle: tokio::runtime::Handle,
-        kanata: KanataClient,
-        handler: Arc<Mutex<FocusHandler>>,
-        status_broadcaster: StatusBroadcaster,
-        pause_broadcaster: PauseBroadcaster,
-        restart_handle: RestartHandle,
-        shutdown_handle: ShutdownHandle,
-        indicator_focus_only: Option<TrayFocusOnly>,
-        retry_delay: Duration,
-        control_builder: B,
-        daemon_bus_name: String,
-    ) -> Self
-    where
-        B: Fn(
-                SniControlMode,
-                tokio::runtime::Handle,
-                KanataClient,
-                Arc<Mutex<FocusHandler>>,
-                StatusBroadcaster,
-                PauseBroadcaster,
-                RestartHandle,
-                ShutdownHandle,
-                Environment,
-                String,
-            ) -> BFut
-            + Send
-            + Sync
-            + 'static,
-        BFut: std::future::Future<Output = Option<SniControl>> + Send + 'static,
-    {
-        let shared_handle: Arc<Mutex<Option<SniIndicatorRuntimeHandle>>> =
-            Arc::new(Mutex::new(None));
-        let task_handle_store = shared_handle.clone();
-        let mut env_receiver = runtime_environment.subscribe();
-        let task = tokio::spawn(async move {
-            let mut active_mode: Option<SniControlMode> = None;
-            let mut active_env: Option<Environment> = None;
-            loop {
-                let env = *env_receiver.borrow();
-                match plan_sni_runtime_transition(active_mode, active_env, env) {
-                    SniRuntimeTransitionPlan::Keep => {}
-                    SniRuntimeTransitionPlan::Stop => {
-                        if let Some(handle) = task_handle_store.lock().unwrap().take() {
-                            println!("[SNI] Shutting down indicator");
-                            drop(handle);
-                        }
-                        active_mode = None;
-                        active_env = None;
-                    }
-                    SniRuntimeTransitionPlan::Start(mode)
-                    | SniRuntimeTransitionPlan::Restart(mode) => {
-                        if let Some(handle) = task_handle_store.lock().unwrap().take() {
-                            println!("[SNI] Shutting down indicator");
-                            drop(handle);
-                        }
-                        active_mode = None;
-                        active_env = None;
-                        let control = control_builder(
-                            mode,
-                            runtime_handle.clone(),
-                            kanata.clone(),
-                            handler.clone(),
-                            status_broadcaster.clone(),
-                            pause_broadcaster.clone(),
-                            restart_handle.clone(),
-                            shutdown_handle.clone(),
-                            env,
-                            daemon_bus_name.clone(),
-                        )
-                        .await;
-                        if let Some(control) = control {
-                            let handle = start_sni_indicator(
-                                control,
-                                status_broadcaster.clone(),
-                                pause_broadcaster.clone(),
-                                indicator_focus_only,
-                            );
-                            *task_handle_store.lock().unwrap() = handle;
-                            active_mode = Some(mode);
-                            active_env = Some(env);
-                        } else {
-                            eprintln!(
-                                "[SNI] Failed to initialize {:?} control; retrying in {}ms unless environment changes",
-                                mode,
-                                retry_delay.as_millis()
-                            );
-                        }
-                    }
-                }
-
-                let should_retry_start =
-                    active_mode.is_none() && sni_control_mode_for_environment(env).is_some();
-                if matches!(
-                    wait_for_sni_runtime_wake_with_delay(
-                        &mut env_receiver,
-                        should_retry_start,
-                        retry_delay,
-                    )
-                    .await,
-                    SniRuntimeWakeReason::ChannelClosed
-                ) {
-                    break;
-                }
-            }
-        });
-        Self {
-            handle: shared_handle,
-            task: Some(task),
-        }
-    }
-}
-
-impl Drop for SniGuard {
-    fn drop(&mut self) {
-        if let Some(task) = self.task.take() {
-            task.abort();
-        }
-        if let Some(handle) = self.handle.lock().unwrap().take() {
-            println!("[SNI] Shutting down indicator");
-            drop(handle);
-        }
-    }
-}
-
-fn dconf_get_bool(key: &str) -> Result<bool, String> {
-    let output = Command::new("dconf")
-        .args(["read", key])
-        .output()
-        .map_err(|error| format!("dconf read failed: {}", error))?;
-
-    if !output.status.success() {
-        let stderr = String::from_utf8_lossy(&output.stderr);
-        return Err(format!("dconf read failed: {}", stderr.trim()));
-    }
-
-    let stdout = String::from_utf8_lossy(&output.stdout);
-    match stdout.trim() {
-        "true" => Ok(true),
-        "false" => Ok(false),
-        "" => Err("key not set".to_string()),
-        value => Err(format!("unexpected dconf output: {}", value)),
-    }
-}
-
-fn dconf_set_bool(key: &str, value: bool) -> Result<(), String> {
-    let value_str = if value { "true" } else { "false" };
-    let output = Command::new("dconf")
-        .args(["write", key, value_str])
-        .output()
-        .map_err(|error| format!("dconf write failed: {}", error))?;
-
-    if !output.status.success() {
-        let stderr = String::from_utf8_lossy(&output.stderr);
-        return Err(format!("dconf write failed: {}", stderr.trim()));
-    }
-
-    Ok(())
-}
-
-fn is_dconf_unavailable(error: &str) -> bool {
-    let lower = error.to_lowercase();
-    lower.contains("no such file or directory")
-        || lower.contains("not found")
-        || lower.contains("failed to execute")
-}
 
 // === GNOME Extension Management ===
